// src/app/feature/model/family-info.model.ts
export interface FamilyInfo {
    id: number;
    userId: number;
    employeeId: number;
    familyMemberName: string;
    relationshipId: number;
    genderId: number;
    dob:string;
    nationalityId:number;
    contactNumber:string;
    profession: string;
    address: string;
    emergencyContact: string;

  }